aux/cleaning_enrichment/requirements.txt
contractions==0.1.73
gensim==4.3.2
Jinja2==3.1.3
langdetect==1.0.9
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
sentence_transformers==2.6.1
spacy==3.7.4
spacy_download==1.1.0
spacy_llm==0.7.1
torch==2.2.2
typer==0.9.4
wasabi==1.1.2
pyarrow==15.0.2
scipy==1.12
#fasttext
fasttext-wheel==0.9.2


